requirements.txt
PyGObject>=3.42.0
python-magic>=0.4.27
requests>=2.31.0
